% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{debugGetDS}
\alias{debugGetDS}
\title{Get DataSHIELD Object}
\usage{
debugGetDS(object)
}
\arguments{
\item{object}{The object to be returned}
}
\value{
The input object, returned to the client side
}
\description{
Returns the object to the client side for debugging purposes
}
