% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{debugHeadDS}
\alias{debugHeadDS}
\title{Get Head of DataSHIELD Object}
\usage{
debugHeadDS(object, n = 6)
}
\arguments{
\item{object}{The dataframe to get the head from}

\item{n}{The number of rows to return (default: 6)}
}
\value{
The first n rows of the input dataframe
}
\description{
Returns the first n rows of a dataframe for debugging purposes
}
