% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{debugTailDS}
\alias{debugTailDS}
\title{Get Tail of DataSHIELD Object}
\usage{
debugTailDS(object, n = 6)
}
\arguments{
\item{object}{The dataframe to get the tail from}

\item{n}{The number of rows to return (default: 6)}
}
\value{
The last n rows of the input dataframe
}
\description{
Returns the last n rows of a dataframe for debugging purposes
}
